refactor: builder canvas
